refactor: remove dummy data
