+++ src/kayenta/canary.dataSource.ts
@@ -11,7 +11,7 @@ import { getCanaryConfigSummaries, listJudges } from './service/canaryConfig.ser
 import { ICanaryConfigSummary, IJudge } from './domain/index';
 import { canaryStore } from './canary';
 import * as Creators from './actions/creators';
-import { listCanaryExecutions } from './service/run/canaryRun.service';
+import { listCanaryExecutions } from './service/canaryRun.service';
 import { ICanaryExecutionStatusResult } from './domain/ICanaryExecutionStatusResult';
 
 export const CANARY_DATA_SOURCE = 'spinnaker.kayenta.canary.dataSource';
