--- src/kayenta/scratch/atlas_canary_config.json
@@ -1,60 +1,0 @@
-{
-  "id": "1",
-  "createdTimestamp": 1498168016035,
-  "updatedTimestamp": 1498168016035,
-  "createdTimestampIso": "2017-06-22T21:46:56.035Z",
-  "updatedTimestampIso": "2017-06-22T21:46:56.035Z",
-  "name": "MySampleAtlasCanaryConfig",
-  "description": "Example Automated Canary Analysis (ACA) Configuration using Atlas",
-  "configVersion": "1.0",
-  "judge": {
-    "judgeConfigurations": {},
-    "name": "NetflixACAJudge-v1.0"
-  },
-  "metrics": [
-    {
-      "name": "cpu",
-      "query": {
-        "type": "atlas",
-        "q": "name,randomValue,:eq"
-      },
-      "groups": [
-        "system"
-      ],
-      "analysisConfigurations": {
-        "canary": {}
-      }
-    },
-    {
-      "name": "requests",
-      "query": {
-        "type": "atlas",
-        "q": "name,apache.http.requests,:eq,:sum"
-      },
-      "groups": [
-        "requests"
-      ],
-      "analysisConfigurations": {
-        "canary": {}
-      }
-    }
-  ],
-  "services": {
-    "atlas": {
-      "name": "atlas",
-      "type": "atlas",
-      "region": "us-east-1",
-      "environment": "prod"
-    }
-  },
-  "classifier": {
-    "groupWeights": {
-      "requests": 50,
-      "system": 50
-    },
-    "scoreThresholds": {
-      "pass": 95,
-      "marginal": 75
-    }
-  }
-}
