--- src/kayenta/service/run/canaryRunStages.ts
@@ -1,2 +1,0 @@
-export const CANARY_JUDGE = 'canaryJudge';
-export const SETUP_CANARY = 'setupCanary';
