--- test/helpers/custom-matchers.d.ts
@@ -1,7 +1,0 @@
-///<reference path="../../node_modules/@types/jasmine/index.d.ts" />
-
-declare namespace jasmine {
-  interface Matchers {
-    textMatch(match: string): boolean;
-  }
-}
