--- test/helpers/customMatchers.ts
@@ -1,24 +1,0 @@
-beforeEach(() => {
-  jasmine.addMatchers({
-    textMatch: () => {
-      return {
-        compare: function (node: JQuery, expected: string) {
-          let actual: string;
-          if (expected === undefined) {
-            expected = '';
-          }
-          if (node !== undefined) {
-            actual = node.text().trim().replace(/\s+/g, ' ');
-          }
-          const result: jasmine.CustomMatcherResult = {pass: expected === actual};
-          if (result.pass) {
-            result.message = `Expected ${expected}`;
-          } else {
-            result.message = `Expected ${expected} but was ${actual}`;
-          }
-          return result;
-        }
-      };
-    }
-  });
-});
