+++ webpack.common.js
@@ -128,11 +128,9 @@ function configure(IS_TEST) {
         FIAT_ENABLED: false,
         INFRA_STAGES: false,
         TIMEZONE: 'America/Los_Angeles',
-        LIVE_CALLS: false,
         METRIC_STORE: 'atlas',
         ATLAS_WEB_COMPONENTS_URL: '',
         REDUX_LOGGER: false,
-        CANARY_REPORTS_ENABLED: false,
         CANARY_STAGES_ENABLED: true,
         TEMPLATES_ENABLED: false,
       }),
